correcao no tipo de movimentacao do montro segundo o tipo de monstro
