+++ src/br/unicamp/mc322/lab10/projeto/mapObjects/characters/heroes/classes/Elf.java
@@ -11,10 +11,10 @@ public class Elf extends Hero {
 	private static final char SPRITE = 'E';
 	private static final String NAME = "Elf";
 	private static final GameTypeObjects ID = GameTypeObjects.ELF;
-	private static final int HP = 3;
-	private static final int MP = 3;
-	private static final int ATTACK_DICES = 1;
-	private static final int DEFENSE_DICES = 1;
+	private static final int HP = 6;
+	private static final int MP = 4;
+	private static final int ATTACK_DICES = 2;
+	private static final int DEFENSE_DICES = 2;
 	
 	public Elf() {
 		super(NAME,ID,HP,MP,new Sprite(SPRITE),ATTACK_DICES,DEFENSE_DICES);
